refactor: Co-locate and reduce editor and toolbar styles
Utilize Gutenberg UI components to fix style bugs.
