+++ src/components/editor/style.scss
@@ -1,0 +1,12 @@
+.gutenberg-kit-editor__toolbar {
+	align-items: center;
+	bottom: 0;
+	left: 0;
+	overflow-x: auto;
+	position: fixed;
+	right: 0;
+}
+
+.gutenberg-kit-editor__toolbar .block-editor-block-contextual-toolbar {
+	width: auto;
+}
